move all components to individual files
