+++ src/components/form/range/_dual_range.scss
@@ -5,23 +5,3 @@
     }
   }
 }
-
-.euiRangeThumb {
-  @include euiCustomControl($type: 'round');
-  @include euiRangeThumbStyle;
-  content: '';
-  position: absolute;
-  left: 0;
-  top: 50%;
-  margin-top: -($euiRangeThumbHeight / 2);
-  z-index: $euiZLevel2;
-  pointer-events: none;
-
-  &:focus {
-    @include euiCustomControlFocused;
-  }
-
-  &--hasTicks {
-    top: 25%;
-  }
-}
