+++ src/components/form/range/_index.scss
@@ -3,7 +3,10 @@
 @import 'range_highlight';
 @import 'range_input';
 @import 'range_label';
+@import 'range_levels';
 @import 'range_slider';
+@import 'range_thumb';
+@import 'range_ticks';
 @import 'range_tooltip';
 @import 'range_track';
 @import 'range_wrapper';
