+++ src/components/form/range/dual_range.js
@@ -9,6 +9,7 @@ import { EuiRangeHighlight } from './range_highlight';
 import { EuiRangeInput } from './range_input';
 import { EuiRangeLabel } from './range_label';
 import { EuiRangeSlider } from './range_slider';
+import { EuiRangeThumb } from './range_thumb';
 import { EuiRangeTrack, LEVEL_COLORS } from './range_track';
 import { EuiRangeWrapper } from './range_wrapper';
 
@@ -402,32 +403,4 @@ EuiDualRange.defaultProps = {
   showRange: true,
   showTicks: false,
   levels: [],
-};
-
-const EuiRangeThumb = ({ min, max, value, disabled, showInput, showTicks, ...rest }) => {
-  const classes = classNames(
-    'euiRangeThumb',
-    {
-      'euiRangeThumb--hasTicks': showTicks
-    },
-  );
-  return (
-    <div
-      className={classes}
-      role="slider"
-      aria-valuemin={min}
-      aria-valuemax={max}
-      aria-valuenow={Number(value)}
-      aria-disabled={!!disabled}
-      tabIndex={(showInput || !!disabled) ? '-1' : '0'}
-      {...rest}
-    />
-  );
-};
-EuiRangeThumb.propTypes = {
-  min: PropTypes.number.isRequired,
-  max: PropTypes.number.isRequired,
-  value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  showInput: PropTypes.bool,
-  showTicks: PropTypes.bool,
 };
