+++ src/components/form/range/range_track.js
@@ -2,9 +2,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-import { range, find } from 'lodash';
+import { range } from 'lodash';
 
-export const LEVEL_COLORS = ['primary', 'success', 'warning', 'danger'];
+import { EuiRangeLevels, LEVEL_COLORS } from './range_levels';
+import { EuiRangeTicks } from './range_ticks';
+
+export { LEVEL_COLORS };
 
 export class EuiRangeTrack extends Component {
 
@@ -115,72 +118,4 @@ EuiRangeTrack.propTypes = {
       color: PropTypes.oneOf(LEVEL_COLORS),
     }),
   ),
-};
-
-const EuiRangeTicks = ({ disabled, onChange, ticks, tickObject, value, max }) => {
-  // Align with item labels across the range by adding
-  // left and right negative margins that is half of the tick marks
-  const ticksStyle = !!ticks ? undefined : { margin: `0 ${tickObject.percentageWidth / -2}%`, left: 0, right: 0 };
-
-  return (
-    <div className="euiRangeTicks" style={ticksStyle}>
-      {tickObject.sequence.map((tickValue) => {
-        const tickStyle = {};
-        let customTick;
-        if (ticks) {
-          customTick = find(ticks, function (o) { return o.value === tickValue; });
-
-          if (customTick == null) {
-            return;
-          } else {
-            tickStyle.left = `${(customTick.value / max) * 100}%`;
-          }
-        } else {
-          tickStyle.width = `${tickObject.percentageWidth}%`;
-        }
-
-        const tickClasses = classNames(
-          'euiRangeTick',
-          {
-            'euiRangeTick--selected': value === tickValue,
-            'euiRangeTick--isCustom': customTick,
-          }
-        );
-
-        return (
-          <button
-            type="button"
-            className={tickClasses}
-            key={tickValue}
-            value={tickValue}
-            disabled={disabled}
-            onClick={onChange}
-            style={tickStyle}
-
-            tabIndex="-1"
-          >
-            {customTick ? customTick.label : tickValue}
-          </button>
-        );
-      })}
-    </div>
-  );
-};
-
-const EuiRangeLevels = ({ levels, max, min, showTicks }) => {
-  const classes = classNames('euiRangeLevels', {
-    'euiRangeLevels--hasTicks': showTicks
-  });
-  return (
-    <div className={classes}>
-      {levels.map((level, index) => {
-        const range = level.max - level.min;
-        const width = (range / (max - min)) * 100;
-
-        return (
-          <span key={index} style={{ width: `${width}%` }} className={`euiRangeLevel euiRangeLevel--${level.color}`} />
-        );
-      })}
-    </div>
-  );
 };
